Fix the vsync test program.
What should I say? Wow. Kernel VSync timestamps are really exact.
Time difference between the vsyncs is constant at 19950 ~ 20050
Next to that, we take the time in userspace, which is more inexact.
What is missing, is the implementation of the DISP_FB_WAITFORVSYNC ioctl
and what time we get from this one. But - the most important thing:
We now have a base to use in libvdpau for exact timing. Hopefully.

Let's go on and finish libvdpau-sunxi :p

Remember that you need a patched kernel for this. I'll provide
a patch soon withing this repository. In the meantime, feel free to discover
https://github.com/rellla/linux-sunxi -> stage/sunxi-3.4_fixes-vsync branch